shrink `t` field
it is only used for a loop `while t < nGroups`, and `nGroups` is in the range `2..=6`, so a `u8` is sufficient

+++ libbz2-rs-sys/src/decompress.rs
@@ -1177,7 +1177,7 @@ pub(crate) fn decompress(
                     error!(BZ_DATA_ERROR);
                 }
                 _ => {
-                    if t < nGroups as i32 {
+                    if t < nGroups {
                         current_block = BZ_X_CODING_1;
                         continue;
                     }
